requirements.txt
openai>=1.0.0
langgraph>=0.1.0
langsmith>=0.1.0
pyyaml>=6.0
rendercv>=1.0.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0 